requirements.txt
flask>=2.0.0
flask-cors>=4.0.0
pillow>=9.0.0
numpy>=1.21.0

